Initial Rewrite
What has changed since the last commit?
The entire codebase has been rewritten. The communication repo is now
utilised. We no longer use rerun. Dependencies are now atomic and use
git rather then relative filepaths.

Why did you choose to make said changes?
The old codebase was rushed due to time constraints and did not provide
adequete abstraction. Rerun is no longer in use due to it being buggy
and slow in certain cases as well as the janky network setup.
Dependencies are now atomic and use git to more easily preserve a
working state and also test with older versions of the deps. The rewrite
also serves as a starting point for designing the API for the upcoming
smartport communication implementation.

Did you consider any alternatives?
An alternative that was considered was to modify the existing code
rather then rewrite the entire codebase. This was decided against due to
the heavy modifications needed anyway to incorporate communcation and
the upcoming smartport communication. It was concluded that it would be
easier and result in better code if there was a rewrite instead.

Did you do any testing at if so how did you test and what were the results?
There was a test run where the main function only contained medatior, a
log call and a 5ms sleep. This was run on the pi and succeeded in
communicating with the GUI remotely. A later test was run to test the
mediator and the robot correctly connecting to the brain at the same
time. This failed due to some incorrect assumptions when creating the
brain which caused a large amount of logs which in turn caused the
remote logging to disconnect. The mediator was swapped out for
SimpleLogger which helped track the issue with creating the connection
with the brain (with heavy assistance from serxka). The test was run
after these changes and it appeared that the code mostly worked. There
was a minor bug where state changes would get reported even if the
change was to the same state which was solved. Another test was done
where controller inputs were printing from the driver function
(controller.ly() and .rx()) which appeared to be working fine.

What challenges did you run into?
There were problems regarding connecting to the brain as well as logging
as mentioned above. Also note that SimpleLogger was swapped back out for
mediator after a commit in communication which logs to both at the same
time. Do note that this has not been testing on the robot itself due to
no access to the robot.

What design changes did you make mid commit?
A lot of the code that now resides in Brain::init() was moved into
Brain::init() from Robot::new() due to there being no reason for it to
be in Robot::new(). The Brain was changed to update stuff directly
rather then returning a pair of packets and having the Robot update
stuff. This was found to make the code neater without exposing
unnecessary data. Changes to logging were made as mentioned above. The
main loop was moved from main into Robot::run().

Who else was involved in this commit?
serxka was a major help in debugging the issues with connecting with the brain.

Did you take inspiration from anywhere or use anything as a reference?
Not in particular.

What challenges still remain/what code still needs to be written?
Testing, testing and more testing. The code is also very minimal in
terms of functionality and does not contain any examples of mechanism
specific code. RobotState::progress() needs to be completed and tested.

List what has been changed in every file
- A large amount of files were removed as it is a rewrite.
- Cargo.lock is autogenerated.
- Cargo.toml now uses git dependencies instead of relative paths and now
  includes the communication repo.
- brain.rs contains all the code for interfacing with the brain as well
  as some helper types: Packet which wraps around StatusPkt with a
  timestamp; The enum State which represents the possible states of
  the brain as calculated by the bitfield in StatusPkt. Brain::init()
  blocks untill it has connected to the brain returns a Brain and Controller
  struct. Brain::update_state will update RobotState and Controller as
  well as update the values instead of each Motor. Brain::write_changes
  will send the target in each Motor to the brain.
- drivebase.rs contains Tankdrive struct which is a wrapper for some
  motors with a method set_side_percent_voltage which is useful for
  controlling how a drivebase moves assuming the motors are in a
  tankdrive configuration. The Tankdrive struct also has a const generic
  for the number of motors on each side of the drivebase.
- controller.rs contains the Controller struct which can be created from
  [Packet; 2] (see brain.rs). The controller struct maps the axes inputs
  from -127..=127 to [-1.0, 1.0] and stores it. The Constroller struct
  also has methods to query if a particular buttons has is held or has
  just been pressed or released. Note that a warning is thrown if you
  try query more then one (or zero) buttons at the same time since it
  was deemed that this is note likely to be intended as probably is a
  bug. This might change in the future if the need arises.
- motor.rs is the struct that represents each possible motor that is
  connected to the brain. Option<MotorState> indicates that the motor
  was is not connected. Controlling the motor can be achieved by writing
  to the target field of robot then calling Brain::write_changes() as
  the brain holds an instance of each motor. The target field contains
  the Target enum which various different targets that the user may want
  to set such as PercentVoltage (from -1 to 1) or RotationalVelocity
  (in RPM).
- robotb.rs does not contain any code currently as robota.rs was used for
  testing.
- robota.rs contains the Robot struct which is what currently holds all
  the logic for the robot and gets run in main. Robot::new() creates
  the all the needed components: RobotState; Brain; Controller;
  TankDrive<0>; Mediator. Note that TankDrive current has a const
  generic argument of 0 which will change when the correct motors for
  testing are figured out. RobotState represents the possible states
  that the robot can be in and decides what function the robot runs in
  the main loop. RobotState is calculate from the current brain::State,
  the last RobotState as well as if the robot is currently in skills
  mode (which will later be decided by a hardware switch but currently
  is a constant). Currently only RobotState::{DriverDriver DriverSkills)
  does anything, both of which trigger the driver() function. The main loop in Robot::run() does the following:
  handle events send to the mediator from the listener thread (see
  communication repo), call Brain::update_state() which updates the
  motors, controller and RobotState then lastly match on RobotState and
  execute the appropriate function (as mentioned before the only
  function implemented is driver()).